Filter for public only constructors
Add a test to cover the scenarios where the no-arg constructor is public, protected and private

Improve the code slightly that finds tries to find an alternative constructor. This could be improved further to find one that matches the schema. Currently it just finds the first public constructor in the array returned by clazz.getConstructors()