Retrieve the no-argument constructor (#1690)
* Retrieve the no-argument constructor using a method called getParameterlessPublicConstructor

This fixes issue #1689 where the no-argument constructor is not always the one you get

* Update JavaObjectTransformer.java

* Filter for public only constructors

Add a test to cover the scenarios where the no-arg constructor is public, protected and private

Improve the code slightly that finds tries to find an alternative constructor. This could be improved further to find one that matches the schema. Currently it just finds the first public constructor in the array returned by clazz.getConstructors()

---------

Co-authored-by: Matthew Tatum <tatum.matthew@gmail.com>
Co-authored-by: Rick M <kingthorin@users.noreply.github.com>